rename: 푸터 컴포넌트 위치 이동 (#34)
