+++ src/shared/ui/index.ts
@@ -1,5 +1,5 @@
 import { Toaster } from '@/shared/ui/Toaster'
-import Footer from '@/shared/ui/footer/Footer'
+import Footer from '@/shared/ui/Footer'
 import { Modal, ModalClose } from '@/shared/ui/Modal'
 import Header from '@/shared/ui/header/Header'
 import { Button, ButtonVariants } from '@/shared/ui/Button'
